#6 Minor changes at the Elbe with the step unneccessary stepcounter. WaterLevel added to dynamically add the waterLevel.

+++ src/test/java/de/uni_oldenburg/simulation/elbe/models/TidesTest.java
@@ -14,7 +14,7 @@ public class TidesTest {
 	private Tides instance;
 	private long highTidePeriod = 25000;
 	private long lowTidePeriod = 20000;
-	private long elbeLength = 507 + 230 + 230 + 200 + 48;
+	private int elbeLength = 507 + 230 + 230 + 200 + 48;
 
 	@Before
 	public void setUp() throws Exception {
